rest-api-agent: Moved the `StackHost` class and made the `port` filed a `String`.
